Selectively skip system VPD if Power is ON
This commit selectively skips just the prime inventory portion
of the parser when collecting system VPD with the CEC powered ON.

We used to skip all of the VPD collection for the system VPD when
powered ON, but it appears that we are losing system VPD on certain
BMC reboots.

This commit works around that bug.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>
Change-Id: Iefce9807275339d8bbdccfb5cb36f365b369fa29