parser: Publish FRU Number For DIMMs
Make the memory VPD parser publish com.ibm.ipzvd.VINI FN.
This is the FRU number and for DDIMMs it is the same as the part number.
This ensures that the same is published over Redfish as the
SparePartNumber.

Tested:

Made sure that when the DIMM VPD is collected, we see both
com.ibm.ipzvpd.VINI FN as well as
xyz.openbmc_project.Inventory.Decorator.Asset SparePartNumber
show up on D-Bus.

Also verified that the SparePartNumber shows up in the GUI.

Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>